Fix/#211

+++ src/pages/coffee-detail/CoffeeDetailButtons.jsx
@@ -1,41 +1,48 @@
-import { Box, Button, Grid } from '@mui/material';
-import { buttonStyles } from './ButtonStyle';
-import ShareIcon from '@mui/icons-material/Share';
-import NewBtn from 'components/common/new-btn/NewBtn';
-import { theme } from 'styles/themeMuiStyle';
-import { useNavigate } from 'react-router-dom';
-import { applyCoffeechat, deleteCoffeechat } from 'api/api';
-import { useEffect, useState } from 'react';
+import { Box, Button, Grid } from "@mui/material";
+import { buttonStyles } from "./ButtonStyle";
+import ShareIcon from "@mui/icons-material/Share";
+import NewBtn from "components/common/new-btn/NewBtn";
+import { theme } from "styles/themeMuiStyle";
+import { useNavigate } from "react-router-dom";
+import { applyCoffeechat, deleteCoffeechat } from "api/api";
+import { useEffect, useState } from "react";
 
-function CoffeeDetailButtons({ id, host, coffeeChatData, isParticipant, setIsParticipant }) {
+function CoffeeDetailButtons({
+  id,
+  host,
+  coffeeChatData,
+  isParticipant,
+  setIsParticipant,
+}) {
+  console.log(coffeeChatData);
   const navigate = useNavigate();
   const [applyCoffee, setApplyCoffee] = useState({});
-  const isButtonDisables = coffeeChatData.status !== '모집중' || isParticipant;
+  const isButtonDisables = coffeeChatData.status !== "모집중" || isParticipant;
 
   const applyButtonText = () => {
     switch (coffeeChatData.status) {
-      case '모집중':
-        return '신청하기';
-      case '모집종료':
-        return '모집 마감';
-      case '모집완료':
-        return '종료된 커피챗입니다.';
+      case "모집중":
+        return "신청하기";
+      case "모집종료":
+        return "모집 마감";
+      case "모집완료":
+        return "종료된 커피챗입니다.";
       default:
-        return '상태 확인 중';
+        return "상태 확인 중";
     }
   };
 
   const applyForCoffeeChat = async () => {
     try {
       await applyCoffeechat(id, applyCoffee);
-      alert('신청이 완료되었습니다.');
+      alert("신청이 완료되었습니다.");
       setIsParticipant(true);
     } catch (error) {
       if (error.response?.status !== 409) {
-        console.log('신청 중 에러가 발생했습니다.');
+        console.log("신청 중 에러가 발생했습니다.");
         return;
       }
-      alert('이미 신청된 커피챗입니다.');
+      alert("이미 신청된 커피챗입니다.");
       return;
     }
   };
@@ -45,12 +52,23 @@ function CoffeeDetailButtons({ id, host, coffeeChatData, isParticipant, setIsPar
   };
 
   const removeCoffeeChat = async () => {
-    try {
-      await deleteCoffeechat(id, applyCoffee);
-      alert('삭제가 완료되었습니다.');
-      navigate('/coffee');
-    } catch (error) {
-      console.log(error);
+    let confirmation = true; // 기본적으로 삭제를 진행
+
+    // 현재 모집 인원이 1명 이상인 경우에만 확인 창을 표시
+    if (coffeeChatData.currentRecruitCount > 0) {
+      confirmation = window.confirm(
+        "삭제하려는 커피챗에 참여 신청한 회원이 존재합니다. 정말 삭제하시겠습니까?"
+      );
+    }
+
+    if (confirmation) {
+      try {
+        await deleteCoffeechat(id, applyCoffee);
+        alert("삭제가 완료되었습니다.");
+        navigate("/coffee");
+      } catch (error) {
+        console.error("삭제 중 오류가 발생했습니다: ", error);
+      }
     }
   };
 
@@ -71,10 +89,10 @@ function CoffeeDetailButtons({ id, host, coffeeChatData, isParticipant, setIsPar
     navigator.clipboard
       .writeText(url)
       .then(() => {
-        alert('URL이 클립보드에 복사되었습니다.');
+        alert("URL이 클립보드에 복사되었습니다.");
       })
       .catch((err) => {
-        console.error('URL 복사에 실패했습니다.', err);
+        console.error("URL 복사에 실패했습니다.", err);
       });
   };
 
@@ -83,20 +101,26 @@ function CoffeeDetailButtons({ id, host, coffeeChatData, isParticipant, setIsPar
       <Box
         container
         sx={{
-          width: '100%',
-          display: 'flex',
-          justifyContent: 'space-between',
-          gap: '20px',
+          width: "100%",
+          display: "flex",
+          justifyContent: "space-between",
+          gap: "20px",
         }}
       >
-        <Grid item xs={8} sm={8} fullwidth sx={{ display: 'flex', gap: '12px', flexGrow: 1 }}>
+        <Grid
+          item
+          xs={8}
+          sm={8}
+          fullwidth
+          sx={{ display: "flex", gap: "12px", flexGrow: 1 }}
+        >
           {host ? (
             <>
               <NewBtn
                 title="수정하기"
                 styles={{
                   background: theme.palette.primary.main,
-                  color: 'white',
+                  color: "white",
                 }}
                 onClick={editCoffeeChat}
               />
@@ -104,17 +128,19 @@ function CoffeeDetailButtons({ id, host, coffeeChatData, isParticipant, setIsPar
                 title="삭제하기"
                 styles={{
                   background: theme.palette.primary.main,
-                  color: 'white',
+                  color: "white",
                 }}
                 onClick={removeCoffeeChat}
               />
             </>
           ) : (
             <NewBtn
-              title={isParticipant ? '신청완료' : applyButtonText()}
+              title={isParticipant ? "신청완료" : applyButtonText()}
               styles={{
-                background: isButtonDisables ? '#EBEBEB' : theme.palette.primary.main,
-                color: 'white',
+                background: isButtonDisables
+                  ? "#EBEBEB"
+                  : theme.palette.primary.main,
+                color: "white",
               }}
               onClick={applyForCoffeeChat}
               disable={isButtonDisables}
